+++ src/App.js
@@ -1,20 +1,21 @@
-import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import logo from './logo.svg';
+import React from 'react';
+import { BrowserRouter as Router, Routes, Route, Link } from 'react-router-dom'; // Import necessary components from react-router-dom
 import './App.css';
 import Login from './Pages/Login';
 import APQPTable from './Forms/ApqpTable';
+import FeasibilityReviewPage from './Forms/FeasibilityReview';
 
 function App() {
   return (
-    <div>
-      <Router>
-        <Routes>
-          <Route path="/login" element={<Login/>}></Route>
-          <Route path="/apqpForm" element={<APQPTable/>}></Route>
+    <Router> {/* Wrap your application in Router */}
+      <div>
+        <Routes> {/* Define your routes */}
+          <Route path="/login" element={<Login />} />
+          <Route path="/apqpform" element={<APQPTable />} />
+          <Route path="/reviewform" element={<FeasibilityReviewPage />} />
         </Routes>
-      </Router>
-    </div>
-
+      </div>
+    </Router>
   );
 }
